Add a spike to get historical data
The spike consists of three steps:
crawl the data
arrange it
export it
This is currently saving it in JSON format, although it could be
easily changed to a DB for example.

This has been built as a spike so a lot of optimizations in terms of
code can be done. A lot of code can be reused and prettified, which
can be done as a follow-up for this project. Also, in terms of
cleaning the data.

In fact, the `denormalization` part is probably the worst bit of
code I've written in a long time. Feel free to read the message at
the top of the `NameExtensions` file to get a gist of what was going
on. Getting rid of that part would speed up the process by tons,
although it would also add a lot of duplicated data. I think that it
may be a good idea if we go down the road of cleaning the data after
it has been imported, but...

Also, please excuse the English, but the historical data source
managed to mess up the election IDs and so they had duplicated data
which resulted in creating a separate URL/portal for fetching that
information. That resulted in... `FuckedUpParlamentareFetcher`. If
things change...